Issue #23 progress
Added a textbox to submit scanned student ID and check if it is logged in and has admin perms for the matching account.
Added studentid to SQLite Database.
Added code related to the SQLite Database into index.js and app.js.
Added comma to "Welcome Back <%=user%>!" in index.js.
scan.ejs added, but unfinished.